# doc/swagger.yaml
definitions:
...
  pkg_server_router.CreateSchemaRequest:
    properties:
      author:
        type: string
      name:
        type: string
      schema:
        $ref: '#/definitions/schema.JSONSchema'
      sign:
        description: Sign represents whether the schema should be signed by the author.
          Default is false.
        type: boolean
    required:
    - author
    - name
    - schema
    type: object
  pkg_server_router.CreateSchemaResponse:
    properties:
      id:
        type: string
      schema:
        $ref: '#/definitions/schema.VCJSONSchema'
      schemaJwt:
        type: string
    type: object
...
  pkg_server_router.GetSchemaResponse:
    properties:
      schema:
        $ref: '#/definitions/schema.VCJSONSchema'
      schemaJwt:
        type: string
    type: object
  pkg_server_router.GetSchemasResponse:
    properties:
      schemas:
        items:
          $ref: '#/definitions/pkg_server_router.GetSchemaResponse'
        type: array
    type: object
...
  pkg_server_router.VerifySchemaRequest:
    properties:
      schemaJwt:
        type: string
    required:
    - schemaJwt
    type: object
  pkg_server_router.VerifySchemaResponse:
    properties:
      reason:
        type: string
      verified:
        type: boolean
    type: object
...
  schema.JSONSchema:
    additionalProperties: true
    type: object
  schema.VCJSONSchema:
    properties:
      author:
        type: string
      authored:
        type: string
      id:
        type: string
      name:
        type: string
      schema:
        $ref: '#/definitions/schema.JSONSchema'
      type:
        type: string
      version:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: tbd-developer@squareup.com
    name: TBD
    url: https://github.com/tavroi/ssi-service/issues
  description: https://github.com/tavroi/ssi-service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: SSI Service API
  version: "0.1"
paths:
...
  /v1/schemas:
    get:
      consumes:
      - application/json
      description: Get schemas
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_server_router.GetSchemasResponse'
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get Schemas
      tags:
      - SchemaAPI
    put:
      consumes:
      - application/json
      description: Create schema
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pkg_server_router.CreateSchemaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/pkg_server_router.CreateSchemaResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create Schema
      tags:
      - SchemaAPI
  /v1/schemas/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a schema by its ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete Schema
      tags:
      - SchemaAPI
    get:
      consumes:
      - application/json
      description: Get a schema by its ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_server_router.GetSchemaResponse'
        "400":
          description: Bad request
          schema:
            type: string
      summary: Get Schema
      tags:
      - SchemaAPI
  /v1/schemas/verification:
    put:
      consumes:
      - application/json
      description: Verify a given schema by its id
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pkg_server_router.VerifySchemaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_server_router.VerifySchemaResponse'
        "400":
          description: Bad request
          schema:
            type: string
      summary: Verify Schema
      tags:
      - SchemaAPI
swagger: "2.0"